refactor(tui): extract rect and layout-tree math into tui::geometry
Pane allocation, split spans, border hit-testing, directional focus
distance, and viewport clamping move out of the module root with their
tests.
